+++ packages/vexide-devices/src/smart/link.rs
@@ -190,6 +190,8 @@ impl RadioLink {
     }
 }
 
+const RADIO_NOT_LINKED: &str = "The radio has not established a link with another radio.";
+
 impl io::Read for RadioLink {
     /// Read some bytes sent to the radio into the specified buffer, returning how many bytes were read.
     ///
@@ -221,22 +223,14 @@ impl io::Read for RadioLink {
     /// ```
     fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
         let is_linked = self.is_linked().map_err(|e| match e {
-            LinkError::Port { source } => match source {
-                PortError::Disconnected => {
-                    io::Error::new(io::ErrorKind::AddrNotAvailable, "Port does not exist.")
-                }
-                PortError::IncorrectDevice => io::Error::new(
-                    io::ErrorKind::AddrInUse,
-                    "Port is in use as another device.",
-                ),
-            },
+            LinkError::Port { source } => source,
             _ => unreachable!(),
         })?;
 
         if !is_linked {
             return Err(io::Error::new(
                 io::ErrorKind::NotConnected,
-                "Radio is not linked!",
+                RADIO_NOT_LINKED,
             ));
         }
 
@@ -277,22 +271,14 @@ impl io::Write for RadioLink {
     /// ```
     fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
         let is_linked = self.is_linked().map_err(|e| match e {
-            LinkError::Port { source } => match source {
-                PortError::Disconnected => {
-                    io::Error::new(io::ErrorKind::AddrNotAvailable, "Port does not exist.")
-                }
-                PortError::IncorrectDevice => io::Error::new(
-                    io::ErrorKind::AddrInUse,
-                    "Port is in use as another device.",
-                ),
-            },
+            LinkError::Port { source } => source,
             _ => unreachable!(),
         })?;
 
         if !is_linked {
             return Err(io::Error::new(
                 io::ErrorKind::NotConnected,
-                "Radio is not linked!",
+                RADIO_NOT_LINKED,
             ));
         }
 
@@ -319,20 +305,12 @@ impl io::Write for RadioLink {
     /// - An error with the kind [`io::ErrorKind::Other`] is returned if the data could not be written to the radio.
     fn flush(&mut self) -> io::Result<()> {
         if !self.is_linked().map_err(|e| match e {
-            LinkError::Port { source } => match source {
-                PortError::Disconnected => {
-                    io::Error::new(io::ErrorKind::AddrNotAvailable, "Port does not exist.")
-                }
-                PortError::IncorrectDevice => io::Error::new(
-                    io::ErrorKind::AddrInUse,
-                    "Port is in use as another device.",
-                ),
-            },
+            LinkError::Port { source } => source,
             _ => unreachable!(),
         })? {
             return Err(io::Error::new(
                 io::ErrorKind::NotConnected,
-                "Radio is not linked!",
+                RADIO_NOT_LINKED,
             ));
         }
 
@@ -396,7 +374,7 @@ pub enum LinkError {
     },
 
     /// Generic port related error.
-    #[snafu(display("{source}"), context(false))]
+    #[snafu(transparent)]
     Port {
         /// The source of the error.
         source: PortError,