+++ packages/vexide-devices/src/smart/mod.rs
@@ -31,7 +31,17 @@
 //!
 //! [`peripherals`]: crate::peripherals
 
+use core::fmt;
+
+use vex_sdk::{
+    vexDeviceGetByIndex, vexDeviceGetStatus, vexDeviceGetTimestamp, V5_DeviceT, V5_DeviceType,
+    V5_MAX_DEVICE_PORTS,
+};
+
+use crate::PortError;
+
 pub mod distance;
+pub mod electromagnet;
 pub mod expander;
 pub mod gps;
 pub mod imu;
@@ -42,9 +52,10 @@ pub mod rotation;
 pub mod serial;
 pub mod vision;
 
-use core::{fmt, time::Duration};
+use core::time::Duration;
 
 pub use distance::DistanceSensor;
+pub use electromagnet::Electromagnet;
 pub use expander::AdiExpander;
 pub use gps::GpsSensor;
 pub use imu::InertialSensor;
@@ -53,15 +64,13 @@ pub use motor::Motor;
 pub use optical::OpticalSensor;
 pub use rotation::RotationSensor;
 pub use serial::SerialPort;
-use vex_sdk::{
-    vexDeviceGetByIndex, vexDeviceGetStatus, vexDeviceGetTimestamp, V5_DeviceT, V5_DeviceType,
-    V5_MAX_DEVICE_PORTS,
-};
+use snafu::ensure;
 pub use vision::VisionSensor;
 
-use crate::PortError;
+use crate::{DisconnectedSnafu, IncorrectDeviceSnafu};
 
 /// Defines common functionality shared by all Smart Port devices.
+
 pub trait SmartDevice {
     /// The interval at which the V5 brain reads packets from Smart devices.
     const UPDATE_INTERVAL: Duration = Duration::from_millis(10);
@@ -149,13 +158,18 @@ pub(crate) fn validate_port(number: u8, device_type: SmartDeviceType) -> Result<
     };
 
     if let Some(connected_type) = connected_type {
-        if connected_type != device_type {
-            // The connected device doesn't match the requested type.
-            return Err(PortError::IncorrectDevice);
-        }
+        // The connected device must match the requested type.
+        ensure!(
+            connected_type == device_type,
+            IncorrectDeviceSnafu {
+                expected: device_type,
+                actual: connected_type,
+                port: number,
+            }
+        );
     } else {
         // No device is plugged into the port.
-        return Err(PortError::Disconnected);
+        return DisconnectedSnafu { port: number }.fail();
     }
 
     Ok(())
@@ -246,13 +260,18 @@ impl SmartPort {
     /// Returns a [`PortError`] if there is not a device of the specified type in this port.
     pub fn validate_type(&self, device_type: SmartDeviceType) -> Result<(), PortError> {
         if let Some(connected_type) = self.device_type() {
-            if connected_type != device_type {
-                // The connected device doesn't match the requested type.
-                return Err(PortError::IncorrectDevice);
-            }
+            // The connected device must match the requested type.
+            ensure!(
+                connected_type == device_type,
+                IncorrectDeviceSnafu {
+                    expected: device_type,
+                    actual: connected_type,
+                    port: self.number,
+                }
+            );
         } else {
             // No device is plugged into the port.
-            return Err(PortError::Disconnected);
+            return DisconnectedSnafu { port: self.number }.fail();
         }
 
         Ok(())
@@ -287,7 +306,7 @@ pub enum SmartDeviceType {
     AiVision,
 
     /// Workcell Electromagnet
-    Magnet,
+    Electromagnet,
 
     /// CTE Workcell Light Tower
     LightTower,
@@ -328,7 +347,7 @@ impl From<V5_DeviceType> for SmartDeviceType {
             V5_DeviceType::kDeviceTypeVisionSensor => Self::Vision,
             V5_DeviceType::kDeviceTypeAdiSensor => Self::Adi,
             V5_DeviceType::kDeviceTypeOpticalSensor => Self::Optical,
-            V5_DeviceType::kDeviceTypeMagnetSensor => Self::Magnet,
+            V5_DeviceType::kDeviceTypeMagnetSensor => Self::Electromagnet,
             V5_DeviceType::kDeviceTypeGpsSensor => Self::Gps,
             V5_DeviceType::kDeviceTypeLightTowerSensor => Self::LightTower,
             V5_DeviceType::kDeviceTypeArmDevice => Self::Arm,
@@ -348,7 +367,7 @@ impl From<SmartDeviceType> for V5_DeviceType {
             SmartDeviceType::Distance => V5_DeviceType::kDeviceTypeDistanceSensor,
             SmartDeviceType::Vision => V5_DeviceType::kDeviceTypeVisionSensor,
             SmartDeviceType::AiVision => V5_DeviceType::kDeviceTypeAiVisionSensor,
-            SmartDeviceType::Magnet => V5_DeviceType::kDeviceTypeMagnetSensor,
+            SmartDeviceType::Electromagnet => V5_DeviceType::kDeviceTypeMagnetSensor,
             SmartDeviceType::LightTower => V5_DeviceType::kDeviceTypeLightTowerSensor,
             SmartDeviceType::Arm => V5_DeviceType::kDeviceTypeArmDevice,
             SmartDeviceType::Optical => V5_DeviceType::kDeviceTypeOpticalSensor,