+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Before releasing:
 
 - You can now detect controller release occurrences with `ButtonState::is_now_released`.
 - Added support for 5.5W motors with a new constructor (`Motor::new_exp`) and four new getters (`Motor::max_voltage`, `Motor::motor_type`, `Motor::is_v5`, and `Motor::is_exp`) for `Motor`. (#167)
+- Added support for the V5 Workcell Electromagnet smart device. (#176)
 - The conditions upon which functions return errors are now documented. (#155).
 - Implemented the `Copy` trait for `BannerTheme`.
 - Added a getter that retrieves a `Controller`'s identifier. (#189)
@@ -35,6 +36,7 @@ Before releasing:
 - Added the `OpticalSensor::GESTURE_UPDATE_INTERVAL` (50mS) constant (#211).
 - Added a `toggle` method to `AdiDigitalOut` to toggle between level outputs.
 - Added a `SerialPort::set_baud_rate` method for the adjusting baudrate of a generic serial smartport after initialization. (#217)
+- Added fields containing relevant failure information to several error types (#221) (**Breaking Change**)
 
 ### Fixed
 
@@ -58,6 +60,7 @@ Before releasing:
 - Changed the incorrect return types of `AdiSolenoid::is_open` and `AdiSolenoid::is_closed` from `LogicLevel` to `bool`. (#164) (**Breaking Change**)
 - Renamed `Motor::MAX_VOLTAGE` to `Motor::V5_MAX_VOLTAGE` and added `Motor::EXP_MAX_VOLTAGE`. (#167) (**Breaking Change**)
 - Moved the ability to convert Smart devices to `SmartPorts` out of the `SmartDevice` trait and into the devices themselves. (#171) (**Breaking Change**)
+- Renamed `SmartDeviceType::Magnet` to `SmartDeviceType::Electromagnet`. (#176) (**Breaking Change**)
 - Getters and constructors will now create warnings when their return values are not used. (#155)
 - Renamed `OpticalSensor::rgb` to `OpticalSensor::color` and `OpticalSensor::raw` to `OpticalSensor::raw_color` (#179) (**Breaking Change**).
 - Made the following functions infallible: `AdiAccelerometer::sensitivity`, `AdiAccelerometer::max_acceleration`, `AdiPotentiometer::potentiometer_type`, `AdiPotentiometer::max_angle`, `Motor::target`, and `RotationSensor::direction`. (#182) (**Breaking Change**)